requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
DAWG-Python==0.7.2
Django==4.2.4
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
environs==9.5.0
et-xmlfile==1.1.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
magic-filter==1.0.10
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
num2words==0.5.13
openpyxl==3.1.2
packaging==23.1
pdfkit==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3
PyVirtualDisplay==3.0
redis==4.6.0
russian-names==0.1.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
